Refactored location of database config and connection test source code to main src tree.  Also fixed unit tests by adding env var values for MYSQL_* in jest test setup file.
