+++ packages/applications-service-api/__tests__/setup-jest.js
@@ -1,0 +1,7 @@
+// Set MYSQL env variables which are loaded statically via imports
+process.env.MYSQL_DIALECT = 'mysql';
+process.env.MYSQL_USERNAME = 'username';
+process.env.MYSQL_PASSWORD = 'password';
+process.env.MYSQL_DATABASE = 'database';
+process.env.MYSQL_HOST = 'host';
+process.env.MYSQL_PORT = '3306';
